BAU: Add GraalVM gradle plugin and new task
The new task defines some properties for building a native image. The
main class is the main class used by Lambda when running your function.
As we're now using a custom runtime rather than the java runtime we need
to include the java-runtime-interface-client. It's this lib that has the
main class we need to invoke the native image build against.