BAU: Add graalVM dir with required resources
The aws-lambda-java-runtime-interface-client library includes
pre-compiled binaries. One of arm64 and one for amd64. Unfortunately
there's a bug and the arm64 binaries don't work. You can fix this by
re-compiling the library locally and it's being committed here for
posterity. It needs to be included in maven local when building.

This also includes a simple python flask app for mocking a part of the
aws lambda runtime API. This is to allow running of the tracing agent
against the runtime-interface-client.

Finally a Dockerfile for building an environment on an M1 that you can
generate a native image in