BAU: Copy testing libs to implementation
This is horrible. but the native-image build was failing with a null
pointer exception that was impossible to diagnose.

I managed to fix it by adding all the test deps to the compile class
path.

There will probably be something in the config files for the
native-image that referenced some test lib by mistake. This would be
narrowed down with a process of elimination.

This was done to move fasst.