BAU: Refactor persistence to use static table schemas
GraalVM doesn't play nicely with DynamoDB enhanced client. It can't cope
with the dynamic loading of table schemas from bean definitions.
Something to do with anonymous inner classes being created. AWS are
aware and working on it.

A work around is to define static table schemas and use those instead.