BAU: Add config files and access-filter
The access-filter is an attempt to keep as much of the dynamic stuff
from the testing libs out of the config as possible. It's not perfect
though - for example Mockito does something weird with generating
inner-classes which don't get filtered.

The config files are a combination of those generated by the tracing
agent when run against the accesstoken lambda test, and the runtime
client againt the python lamdba mock. They also have a fair amount of
manual fettering added to them too.