BAU: Add a proper event body to lambda mock
This takes an example event from API gateway and adds it to the lambda
mock. This causes the actual classes involved in processing the request
to get run and picked up by the tracing agent. Previously we were just
getting a null input and none of the right reflection config.